requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
flask-ngrok==0.0.25
frozenlist==1.4.1
future==1.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
multidict==6.0.5
ngrok==1.2.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
